Paramterize Client on Key Provider
By parameterizing the client on the key provided, we're able to avoid dynamic dispatch on the
key provider (eg, it's no longer a trait object).

My real motivation here is that I want to be able to add Send to the KeyProvider
trait object in the client, but I can't, because I want to send the client between threads,
but I can't. Simply wrapping the client in an Arc does not solve the problem, the only
fix seems to be annotating the KeyProvider on the client with Send.

However, annotating the KeyProvider with send is more restrictive, and I think the underlying issue is
that we want to parameterize over the KeyProvider rather than use a trait object. This is more flexible
and allows the caller to dictate the bounds of the KeyProvider.

This change also makes public the KeyProvider and GoogleKeyProvider classes. This seems reasonable
to me, given that we already expose the ability to override the KeyProvider (without actually allowing users
to implement their own KeyProvider class, which seems odd, and I may be misunderstanding something).