:sparkles: Implement firebase auth token verification
Abstract GoogleKeyProvider to a trait that is implemented by GoogleSignKeyProvider and FirebaseAuthenticationKeyProvider. Abstract RequiredClaims and IdPayload to associated types of the Validator trait which is implemented by GoogleSigninValidator and FirebaseValidator. Still need to expose the new API, include a check for auth_time claim for FirebaseValidator, expand FirebaseIdPayload, test FirebaseClient using custom key provider and refactor.