:recycle: Refactor Error, ignore expiry -> mock timestamp
A lot of information was being lost when converting 3rd party library errors to this crate's Error type. Error::InvalidToken is particularly unhelpful for users of this library because it represented so many different underlying errors. Error has been restructured so that it contains all underlying error information. Replaced unsafe_ignore_expiration method with unsafe_mock_timestamp method that it can be used to test firebase token validation in a deterministic way. Instead of claims_are_valid method for Validator, use validate_claims method that returns the error corresponding to a failing validation condition of the claim. This allows specific error information to propogate to the library user.